Add files via upload
This contains code for Assignment 2 for Computer Vision. This involved 2 semantic segmentation tasks. The first task consisted of implementation of mean shift algorithm for segmentation of an RGB image. The second task consisted of semantic segmentation of an image comprised of multiple MNIST images. A modified version of SegNet was implemented for the task.